[feat]: [CDS-47025] ASG NG - Make Rolling deploy parameters to be allowed as runtime inputs & change parameter type (#42030)
* [feat]: [CDS-47025] asg ng make rolling deploy parameters to be allowed as runtime input

* [feat]: [CDS-47025] asg ng make minhealthypercentage parameter to integer type

+++ 125-cd-nextgen/src/main/java/io/harness/cdng/aws/asg/AsgRollingDeployStepInfo.java
@@ -52,7 +52,7 @@ public class AsgRollingDeployStepInfo extends AsgRollingDeployBaseStepInfo imple
   @Builder(builderMethodName = "infoBuilder")
   public AsgRollingDeployStepInfo(ParameterField<List<TaskSelectorYaml>> delegateSelectors,
       ParameterField<Boolean> skipMatching, ParameterField<Boolean> useAlreadyRunningInstances,
-      ParameterField<Integer> instanceWarmup, ParameterField<Double> minimumHealthyPercentage) {
+      ParameterField<Integer> instanceWarmup, ParameterField<Integer> minimumHealthyPercentage) {
     super(delegateSelectors, skipMatching, useAlreadyRunningInstances, instanceWarmup, minimumHealthyPercentage);
   }
